Eskills Retrofit Stuff moved to eskills module
