+++ core/network/eskills/src/main/java/com/appcoins/wallet/core/network/eskills/annotations/EskillsRetrofit.kt
@@ -1,0 +1,13 @@
+package com.appcoins.wallet.core.network.eskills.annotations
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+@Target(
+  AnnotationTarget.CLASS,
+  AnnotationTarget.FIELD,
+  AnnotationTarget.FUNCTION,
+  AnnotationTarget.VALUE_PARAMETER
+)
+annotation class EskillsCarouselRetrofit
