+++ core/utils/properties/src/main/java/com/appcoins/wallet/core/utils/properties/HostProperties.kt
@@ -13,7 +13,7 @@ object HostProperties {
   const val APTOIDE_WEB_SERVICES_AB_TEST_HOST = "https://abtest.aptoide.com/api/v1/"
   const val RAKAM_BASE_HOST = "https://rakam-api.aptoide.com"
   const val FEEDBACK_ZENDESK_BASE_HOST = "https://aptoide.zendesk.com/api/v2/"
-  const val ESKILLS_CAROUSEL_HOST = "https://ws2-cache.aptoide.com/api/7/apps/"
+  const val ESKILLS_CAROUSEL_HOST = "https://ws2-cache.aptoide.com/api/7/"
 }
 
 private const val BASE_HOST_PROD = "https://api.catappult.io"
